Use ceill instead of ceil

+++ src/timeout.c
@@ -177,7 +177,7 @@ int getTimeoutFromObjectOrReply(client *c, robj *object, mstime_t *timeout, int 
             addReplyError(c, "timeout is out of range");
             return C_ERR;
         }
-        tval = (long long) ceil(ftval);
+        tval = (long long) ceill(ftval);
     } else {
         if (getLongLongFromObjectOrReply(c,object,&tval,
             "timeout is not an integer or out of range") != C_OK)
